Added (and filled in) RemoveCommitmentController and RemoveCOmmitmentRequestObserver.
Tested functionality of removing an event from the database using the
"next" button. Removed this test feature before committing, as it would
be bad if this was forgotten.

Fixed error in address sent to server, changed from
calendar/event/project/ to calendar/event/.

Discovered and fixed spelling errors in database address in
AddCommitmentController and EditCommitmentController